refactor(drawer) change component structure
