+++ src/components/drawer/drawer-context.ts
@@ -1,8 +1,8 @@
 import { createContext } from 'react';
 
-import { DrawerRootProps } from './drawer-root';
+import { DrawerProps } from '.';
 
-export const DrawerContext = createContext<DrawerRootProps>({
+export const DrawerContext = createContext<DrawerProps>({
   open: false,
   onClose: () => null,
   placement: 'left',
