+++ src/components/drawer/drawer.test.tsx
@@ -10,9 +10,9 @@ describe('Drawer', () => {
 
   it('should render correctly', () => {
     render(
-      <Drawer.Root open={false} onClose={() => {}} data-testid="drawer">
+      <Drawer open={false} onClose={() => {}} data-testid="drawer">
         drawer
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawer = screen.getByTestId('drawer');
@@ -21,14 +21,14 @@ describe('Drawer', () => {
 
   it('should render with class name', () => {
     render(
-      <Drawer.Root
+      <Drawer
         className="custom"
         open={false}
         onClose={() => {}}
         data-testid="drawer"
       >
         drawer
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawer = screen.getByTestId('drawer');
@@ -37,9 +37,9 @@ describe('Drawer', () => {
 
   it('should render the open Drawer', () => {
     render(
-      <Drawer.Root open={true} onClose={() => {}} data-testid="drawer">
+      <Drawer open={true} onClose={() => {}} data-testid="drawer">
         drawer
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawer = screen.getByTestId('drawer');
@@ -48,9 +48,9 @@ describe('Drawer', () => {
 
   it('should render the close Drawer', () => {
     render(
-      <Drawer.Root open={false} onClose={() => {}} data-testid="drawer">
+      <Drawer open={false} onClose={() => {}} data-testid="drawer">
         drawer
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawer = screen.getByTestId('drawer');
@@ -59,11 +59,11 @@ describe('Drawer', () => {
 
   it('should open drawer content with open=true', () => {
     render(
-      <Drawer.Root open={true} onClose={() => {}}>
+      <Drawer open={true} onClose={() => {}}>
         <Drawer.Content data-testid="drawer-content">
           <div>drawer content</div>
         </Drawer.Content>
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawerContent = screen.getByTestId('drawer-content');
@@ -72,11 +72,11 @@ describe('Drawer', () => {
 
   it('should not open drawer content with open=false', () => {
     render(
-      <Drawer.Root open={false} onClose={() => {}}>
+      <Drawer open={false} onClose={() => {}}>
         <Drawer.Content data-testid="drawer-content">
           <div>drawer content</div>
         </Drawer.Content>
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawerContent = screen.getByTestId('drawer-content');
@@ -85,14 +85,14 @@ describe('Drawer', () => {
 
   it('should render drawer with provided placement', () => {
     render(
-      <Drawer.Root
+      <Drawer
         open={false}
         placement="bottom"
         onClose={() => {}}
         data-testid="drawer"
       >
         drawer
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawer = screen.getByTestId('drawer');
@@ -101,14 +101,9 @@ describe('Drawer', () => {
 
   it('should render drawer with provided size', () => {
     render(
-      <Drawer.Root
-        open={false}
-        size="full"
-        onClose={() => {}}
-        data-testid="drawer"
-      >
+      <Drawer open={false} size="full" onClose={() => {}} data-testid="drawer">
         drawer
-      </Drawer.Root>,
+      </Drawer>,
     );
 
     const drawer = screen.getByTestId('drawer');
